requirements.txt
flask>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
python-dotenv>=1.0.0
requests>=2.31.0

# Provider SDKs
openai>=1.0.0
google-genai>=0.6.0
groq>=0.9.0

# Development and testing
pytest>=8.0.0
# pip install pytest-cov
